Fixed null pointer on rfc_call when buffer size is not present
When an array is specified, its length must be present. if it is not present, the va_arg function tries to pull its content and discovers the null-terminator, which it interprets as the length being zero. The next va_arg shouldn't execute but the null pointer was skipped, so a fatal memory error occurs.